выносит функцию pluralizeRus из ads.js в отдельный модуль и импортирует её обратно
